stockfish_explain/requirements.txt
CairoSVG
captum
chess
click
matplotlib
numpy
pandas
Pillow
PyYAML
stockfish
torch
pytest
